requirements.txt
argon2-cffi==18.3.0
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
dj-database-url==0.5.0
Django==1.11.3
django-heroku==0.3.1
django-simple-history==2.5.1
djangorestframework==3.8.2
gunicorn==19.9.0
mysqlclient==1.3.13
psycopg2==2.7.6.1
pycparser==2.19
python-decouple==3.1
pytz==2018.5
six==1.11.0
whitenoise==4.1.2
